Move makedir to a more generally usable place
